requirements.txt
python-telegram-bot>=20.0
openai>=1.0.0
python-dotenv>=1.0.0
nest-asyncio>=1.5.0
pathlib>=1.0.1
pydantic>=2.0.0
python-json-logger>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
pytz>=2023.3
python-dateutil>=2.8.2
requests>=2.31.0
aiohttp>=3.8.5
asyncio>=3.4.3
typing-extensions>=4.7.1
fastapi>=0.68.0
uvicorn>=0.15.0
httpx>=0.24.0
python-multipart>=0.0.5 